+++ src/examples/java/io/nats/examples/NatsPub.java
@@ -16,7 +16,6 @@ package io.nats.examples;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 
-import io.nats.client.AuthHandler;
 import io.nats.client.Connection;
 import io.nats.client.Nats;
 import io.nats.client.Options;
@@ -24,10 +23,11 @@ import io.nats.client.Options;
 public class NatsPub {
 
     static final String usageString =
-            "\nUsage: java NatsPub [server] <subject> <text message>"
+            "\nUsage: java NatsPub [server] <subject> <text message>\n"
             + "\nUse tls:// or opentls:// to require tls, via the Default SSLContext\n"
-            + "Set the environment variable NATS_NKEY to use challenge resposne authentication by setting a file containing your seed.\n"
-            + "Use the URL for user/pass/token authentication.\n";
+            + "\nSet the environment variable NATS_NKEY to use challenge resposne authentication by setting a file containing your seed.\n"
+            + "\nSet the environment variable NATS_CREDS to use JWT/NKey authentication by setting a file containing your user creds.\n"
+            + "\nUse the URL for user/pass/token authentication.\n";
 
     public static void main(String args[]) {
         String subject;
@@ -48,17 +48,13 @@ public class NatsPub {
         }
 
         try {
-            
-            Options.Builder builder = new Options.Builder().server(server).noReconnect();
+            Connection nc = Nats.connect(ExampleUtils.createExampleOptions(server, false));
 
-            if (System.getenv("NATS_NKEY") != null) {
-                AuthHandler handler = new ExampleAuthHandler(System.getenv("NATS_NKEY"));
-                builder.authHandler(handler);
-            }
-
-            Connection nc = Nats.connect(builder.build());
+            System.out.println();
+            System.out.printf("Sending %s on %s, server is %s\n", message, subject, server);
+            System.out.println();
             nc.publish(subject, message.getBytes(StandardCharsets.UTF_8));
-            nc.flush(Duration.ZERO);
+            nc.flush(Duration.ofSeconds(5));
             nc.close();
 
         } catch (Exception exp) {