+++ src/examples/java/io/nats/examples/NatsQSub.java
@@ -16,7 +16,6 @@ package io.nats.examples;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 
-import io.nats.client.AuthHandler;
 import io.nats.client.Connection;
 import io.nats.client.Message;
 import io.nats.client.Nats;
@@ -26,10 +25,11 @@ import io.nats.client.Subscription;
 public class NatsQSub {
 
     static final String usageString =
-            "\nUsage: java NatsQSub [server] <subject> <queue> <msgCount>"
+            "\nUsage: java NatsQSub [server] <subject> <queue> <msgCount>\n"
             + "\nUse tls:// or opentls:// to require tls, via the Default SSLContext\n"
-            + "Set the environment variable NATS_NKEY to use challenge resposne authentication by setting a file containing your seed.\n"
-            + "Use the URL for user/pass/token authentication.\n";
+            + "\nSet the environment variable NATS_NKEY to use challenge resposne authentication by setting a file containing your seed.\n"
+            + "\nSet the environment variable NATS_CREDS to use JWT/NKey authentication by setting a file containing your user creds.\n"
+            + "\nUse the URL for user/pass/token authentication.\n";
 
     public static void main(String args[]) {
         String subject;
@@ -53,18 +53,11 @@ public class NatsQSub {
         }
 
         try {
-            
-            Options.Builder builder = new Options.Builder().server(server).noReconnect();
-
-            if (System.getenv("NATS_NKEY") != null) {
-                AuthHandler handler = new ExampleAuthHandler(System.getenv("NATS_NKEY"));
-                builder.authHandler(handler);
-            }
-
-            Connection nc = Nats.connect(builder.build());
+            Connection nc = Nats.connect(ExampleUtils.createExampleOptions(server, true));
             Subscription sub = nc.subscribe(subject, queue);
-            nc.flush(Duration.ZERO);
+            nc.flush(Duration.ofSeconds(5));
 
+            System.out.println();
             for(int i=0;i<msgCount;i++) {
                 Message msg = sub.nextMessage(Duration.ofHours(1));