+++ src/examples/java/io/nats/examples/NatsReq.java
@@ -15,8 +15,8 @@ package io.nats.examples;
 
 import java.nio.charset.StandardCharsets;
 import java.util.concurrent.Future;
+import java.util.concurrent.TimeUnit;
 
-import io.nats.client.AuthHandler;
 import io.nats.client.Connection;
 import io.nats.client.Message;
 import io.nats.client.Nats;
@@ -25,10 +25,11 @@ import io.nats.client.Options;
 public class NatsReq {
 
     static final String usageString =
-            "\nUsage: java NatsReq [server] <subject> <text message>"
+            "\nUsage: java NatsReq [server] <subject> <text message>\n"
             + "\nUse tls:// or opentls:// to require tls, via the Default SSLContext\n"
-            + "Set the environment variable NATS_NKEY to use challenge resposne authentication by setting a file containing your seed.\n"
-            + "Use the URL for user/pass/token authentication.\n";
+            + "\nSet the environment variable NATS_NKEY to use challenge resposne authentication by setting a file containing your seed.\n"
+            + "\nSet the environment variable NATS_CREDS to use JWT/NKey authentication by setting a file containing your user creds.\n"
+            + "\nUse the URL for user/pass/token authentication.\n";
 
     public static void main(String args[]) {
         String subject;
@@ -49,21 +50,15 @@ public class NatsReq {
         }
 
         try {
-            
-            Options.Builder builder = new Options.Builder().server(server).noReconnect();
-
-            if (System.getenv("NATS_NKEY") != null) {
-                AuthHandler handler = new ExampleAuthHandler(System.getenv("NATS_NKEY"));
-                builder.authHandler(handler);
-            }
-
-            Connection nc = Nats.connect(builder.build());
+            Connection nc = Nats.connect(ExampleUtils.createExampleOptions(server, false));
             Future<Message> replyFuture = nc.request(subject, message.getBytes(StandardCharsets.UTF_8));
-            Message reply = replyFuture.get();
+            Message reply = replyFuture.get(5, TimeUnit.SECONDS);
 
+            System.out.println();
             System.out.printf("Received reply \"%s\" on subject \"%s\"\n", 
                                     new String(reply.getData(), StandardCharsets.UTF_8), 
                                     reply.getSubject());
+            System.out.println();
 
             nc.close();