+++ src/main/java/io/nats/client/impl/NatsConnection.java
@@ -14,13 +14,11 @@
 package io.nats.client.impl;
 
 import java.io.IOException;
-import java.net.URI;
 import java.nio.ByteBuffer;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashSet;
@@ -94,7 +92,6 @@ class NatsConnection implements Connection {
     private CompletableFuture<DataPort> dataPortFuture;
     private DataPort dataPort;
     private String currentServerURI;
-    private String connectServerURI;
     private CompletableFuture<Boolean> reconnectWaiter;
 
     private NatsConnectionReader reader;
@@ -171,7 +168,6 @@ class NatsConnection implements Connection {
             tryToConnect(serverURI);
 
             if (isConnected()) {
-                connectServerURI = serverURI;
                 break;
             } else {
                 updateStatus(Status.DISCONNECTED);
@@ -275,7 +271,6 @@ class NatsConnection implements Connection {
     // will wait for any previous attempt to complete, using the reader.stop and
     // writer.stop
     void tryToConnect(String serverURI) {
-        System.out.println("##### "+serverURI);
         try {
             statusLock.lock();
             try {
@@ -368,7 +363,6 @@ class NatsConnection implements Connection {
             processException(exp);
             throw exp;
         } catch (Exception exp) { // every thing else
-            exp.printStackTrace();
             processException(exp);
             try {
                 this.closeSocket(false);
@@ -977,7 +971,7 @@ class NatsConnection implements Connection {
         }
     }
 
-    void sendConnect(String serverURI) {
+    void sendConnect(String serverURI) throws IOException {
         try {
             NatsServerInfo info = this.serverInfo.get();
             StringBuilder connectString = new StringBuilder();
@@ -986,9 +980,11 @@ class NatsConnection implements Connection {
             String connectOptions = this.options.buildProtocolConnectOptionsString(serverURI, info.isAuthRequired(), info.getNonce());
             connectString.append(connectOptions);
             NatsMessage msg = new NatsMessage(connectString.toString());
+            
             queueInternalOutgoing(msg);
         } catch (Exception exp) {
-            processException(exp);
+            exp.printStackTrace();
+            throw new IOException("Error sending connect string", exp);
         }
     }
     
@@ -1284,47 +1280,27 @@ class NatsConnection implements Connection {
 
     public Collection<String> getServers() {
         NatsServerInfo info = this.serverInfo.get();
-        HashSet<String> servers = new HashSet<String>();
+        HashSet<String> check = new HashSet<String>();
+        ArrayList<String> servers = new ArrayList<>();
 
-        options.getServers().stream().forEach(x -> servers.add(x.toString()));
+        options.getServers().stream().forEach(x -> {
+            String uri = x.toString();
+            if (!check.contains(uri)) {
+                servers.add(uri);
+                check.add(uri);
+            }
+        });
 
         if (info != null && info.getConnectURLs() != null) {
-            servers.addAll(Arrays.asList(info.getConnectURLs()));
+            for (String uri : info.getConnectURLs()) {
+                if (!check.contains(uri)) {
+                    servers.add(uri);
+                    check.add(uri);
+                }
+            }
         }
 
         return servers;
-    }
-
-    public boolean isReconnectServer(String address, int port) {
-        NatsServerInfo info = this.serverInfo.get();
-        if (!this.writer.getReconnectMode()) {
-            return false;
-        }
-
-        if (info == null || info.getConnectURLs() == null) {
-            return false;
-        }
-
-        String hostPort = address+":"+port;
-        String[] connectURLs = info.getConnectURLs();
-
-        try {
-            URI url = this.options.createURIForServer(hostPort);
-            for (int i=0;i<connectURLs.length;i++) {
-                URI connectURI = this.options.createURIForServer(connectURLs[i]);
-                if (connectURI.toString().equals(url.toString())) {
-                    return true;
-                }
-            }
-        } catch (Exception exp) {
-            return false;
-        }
-
-        return false;
-    }
-
-    public String getInitialServerURL() {
-        return this.connectServerURI;
     }
 
     public String getConnectedUrl() {