+++ src/main/java/io/nats/client/impl/SocketDataPort.java
@@ -19,10 +19,7 @@ import java.io.OutputStream;
 import java.net.InetSocketAddress;
 import java.net.Socket;
 import java.net.URI;
-import java.security.cert.Certificate;
 import java.time.Duration;
-import java.util.Arrays;
-import java.util.HashSet;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
 
@@ -42,7 +39,6 @@ public class SocketDataPort implements DataPort {
 
     private InputStream in;
     private OutputStream out;
-    private Certificate[] certificatesFromConnection;
 
     public void connect(String serverURI, NatsConnection conn) throws IOException {
 
@@ -71,20 +67,11 @@ public class SocketDataPort implements DataPort {
      * If the data port type doesn't support SSL it should throw an exception.
      */
     public void upgradeToSecure() throws IOException {
-        boolean recheckReconnect = this.connection.getOptions().recheckClusterCert();
-        boolean isReconnectServer = this.connection.isReconnectServer(this.host, this.port);
-        SSLContext context = null;
-        
-        if (isReconnectServer && !recheckReconnect) {
-            HashSet<Certificate> certs = new HashSet<>();
-            certs.addAll(Arrays.asList(this.certificatesFromConnection));
-            context = SSLUtils.createReconnectContext(certs);
-        } else {
-            context = this.connection.getOptions().getSslContext();
-        }
+        Options options = this.connection.getOptions();
+        SSLContext context = options.getSslContext();
         
         SSLSocketFactory factory = context.getSocketFactory();
-        Duration timeout = this.connection.getOptions().getConnectionTimeout();
+        Duration timeout = options.getConnectionTimeout();
 
         this.sslSocket = (SSLSocket) factory.createSocket(socket, null, true);
         this.sslSocket.setUseClientMode(true);
@@ -103,12 +90,6 @@ public class SocketDataPort implements DataPort {
             this.connection.handleCommunicationIssue(ex);
         }
 
-        // Reset the allowed certs on a full reconnect
-        if (!isReconnectServer || recheckReconnect) {
-            this.certificatesFromConnection = this.sslSocket.getSession().getPeerCertificates();
-        }
-
-        //in = Channels.newChannel(new BufferedInputStream(sslSocket.getInputStream()));
         in = sslSocket.getInputStream();
         out = sslSocket.getOutputStream();
     }