+++ src/test/java/io/nats/client/impl/ConnectionListenerTests.java
@@ -50,7 +50,6 @@ public class ConnectionListenerTests {
             try {
                 assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
                 assertEquals(ts.getURI(), nc.getConnectedUrl());
-                assertEquals(ts.getURI(), ((NatsConnection)nc).getInitialServerURL());
             } finally {
                 nc.close();
                 assertTrue("Closed Status", Connection.Status.CLOSED == nc.getStatus());